extract css regex functions from html
